--- SistemaTransporte/src/modelo/Pessoa.java
@@ -1,44 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package modelo;
-
-/**
- *
- * @author Fernando
- */
-
-public abstract class Pessoa {
-    private String nome;
-    private int idade;
-
-    public Pessoa(String nome, int idade) {
-        this.nome = nome;
-        this.idade = idade;
-    }
-
-    // getters e setters
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public int getIdade() {
-        return idade;
-    }
-
-    public void setIdade(int idade) {
-        this.idade = idade;
-    }
-
-    // metodo para exibir as informacoes da Pessoa
-    public void exibirInfo() {
-        System.out.println("Nome: " + nome);
-        System.out.println("Idade: " + idade + " anos");
-    }
-}
-
